fix: fe refactoring
